requirements.txt
beautifulsoup4
flask
Flask-Cors
requests
flask_caching
redis
python-dotenv